+++ Localization/itIT.Plugins.lua
@@ -4,7 +4,7 @@ local L = addon.Localization.GetDefaultLocale();
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-11-20 19-31-00 ]] --
+-- [[ Exported at 2024-10-27 12-33-20 ]] --
 L["Can I Mog It"] = true
 L["Can I Mog It Desc"] = [=[Questo plugin corregge la sovrapposizione di icone negli oggetti dei mercanti quando sono applicati diversi filtri.