+++ Localization/Localization.lua
@@ -3,5 +3,5 @@ addon.Localization = {};
 local localization = addon.Localization;
 
 function localization.GetDefaultLocale()
-    return LibStub(addon.Libs.AceLocale):NewLocale(addonName, "enUS", true, true);
+    return LibStub(addon.Libs.AceLocale):NewLocale(addonName, "enUS", true);
 end